Renamed current_sense -> load_switch, bts_7040_7008->bts_7040, updated dependencies
