+++ libraries/ms-drivers/rules.mk
@@ -13,5 +13,5 @@ ifeq (x86,$(PLATFORM))
 $(T)_test_ltc_afe_MOCKS := spi_exchange
 $(T)_test_ads1259_adc_MOCKS := spi_exchange
 $(T)_test_adt7476a_fan_controller_MOCKS := i2c_write i2c_read_reg
-$(T)_test_bts_7200_current_sense_MOCKS := adc_read_converted
+$(T)_test_bts_7200_load_switch_MOCKS := adc_read_converted
 endif
